+++ R/cohorts.R
@@ -481,6 +481,8 @@ generatePixelGroups <- function(pixelDataTable, maxPixelGroup,
 #' @export
 speciesEcoregionLatestYear <- function(speciesEcoregion, currentTime) {
   spEco <- speciesEcoregion[year <= currentTime]
+  if (getOption("LandR.assertions", TRUE))
+    if (NROW(spEco) == 0) stop("speciesEcoregion does not have data for the current time")
   spEco[year == max(year)]
 }